requirements.txt
python-binance
pytest
unicorn-binance-websocket-api
eth-brownie
waiting
func_timeout
tronpy